requirements.txt
asgiref==3.2.10
astroid==2.4.2
autopep8==1.5.4
certifi==2020.6.20
chardet==3.0.4
cloudinary==1.22.0
colorama==0.4.4
dj-database-url==0.5.0
Django==3.1.2
django-cloudinary-storage==0.3.0
django-debug-toolbar==3.1.1
djangorestframework==3.12.2
gunicorn==20.0.4
idna==2.10
isort==5.6.3
lazy-object-proxy==1.4.3
mccabe==0.6.1
Pillow==7.2.0
psycopg2==2.8.6
pycodestyle==2.6.0
pylint==2.6.0
python-decouple==3.3
pytz==2020.1
requests==2.24.0
six==1.15.0
sqlparse==0.4.1
stripe==2.55.0
toml==0.10.1
urllib3==1.25.11
whitenoise==5.2.0
wrapt==1.12.1
